requirements/development.txt
-r common.txt

coverage
django-debug-toolbar
ipython
django-spaghetti-and-meatballs
django-extensions
selenium

sphinx
sphinx_rtd_theme
pyenchant
sphinxcontrib-spelling
recommonmark

# pygraphviz  # Has issues accessing the graphviz libraries locally